Added font slider and fixes
