+++ shared/budgie-components/src/lib/display-settings/DisplaySettings.tsx
@@ -1,7 +1,9 @@
 'use client';
-
-import styles from './DisplaySettings.module.css';
+import React, { useEffect } from 'react';
+import { useState } from 'react';
 import "../../root.css";
+import styles from './DisplaySettings.module.css';
+
 
 /* eslint-disable-next-line */
 export interface DisplaySettingsProps {
@@ -9,6 +11,29 @@ export interface DisplaySettingsProps {
 }
 
 export function DisplaySettings(props: DisplaySettingsProps) {
+  const [fontSizeMultiplier, setFontSizeMultiplier] = useState(1);
+
+  useEffect(() => {
+    const rootStyles = getComputedStyle(document.documentElement);
+    const fontSizeValue = parseFloat(rootStyles.getPropertyValue('--font-size-multiplier'));
+    setFontSizeMultiplier(fontSizeValue);
+
+    const fontSlider = document.getElementById("fontSlider") as HTMLInputElement;
+    fontSlider.value = fontSizeValue.toString();
+    const handleSliderChange = () => {
+        const fontSizeValue = parseFloat(fontSlider.value);
+        setFontSizeMultiplier(fontSizeValue);
+        document.documentElement.style.setProperty('--font-size-multiplier', fontSizeValue.toString());
+    };
+
+
+    fontSlider.addEventListener("input", handleSliderChange);
+
+    return () => {
+        fontSlider.removeEventListener("input", handleSliderChange);
+    };
+}, []);
+
     const setLight = () => {
         document.documentElement.setAttribute('data-theme', 'light');
     };
@@ -30,12 +55,15 @@ export function DisplaySettings(props: DisplaySettingsProps) {
         </div>
         <div className={styles.settingsOptionsContainer}>
         <div className={styles.settingsOption}>
-          <p className={styles.settingTitle}>Font Size</p>
-          <p className={styles.settingDescription}>Edit your profile information or change your password.</p>
+          <p className={styles.settingTitle}>Font Size</p>        
+          <div className={styles.sliderContainer}>
+              <span style={{marginRight: "1rem", fontSize: "1rem" }}>A</span>
+              <input type="range" id="fontSlider" min="1" max="2" step="0.1" style ={{width: "15rem"}}></input>
+              <span style={{ marginLeft: "1rem", fontSize: "2rem" }}>A</span>
+          </div>
         </div>
         <div className={styles.settingsOption}>
           <p className={styles.settingTitle}>Colour</p>
-          <p className={styles.settingDescription}>Change the website’s font size, colour, and background.</p>
         </div>
         <div className={styles.settingsOption}>
           <p className={styles.settingTitle}>Background</p>
